+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/CameraXProxy.java
@@ -13,6 +13,7 @@ import androidx.camera.core.CameraSelector;
 import androidx.camera.core.ImageAnalysis;
 import androidx.camera.core.ImageCapture;
 import androidx.camera.core.Preview;
+import androidx.camera.video.Recorder;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugins.camerax.GeneratedCameraXLibrary.ResolutionInfo;
 import java.io.File;
@@ -23,45 +24,57 @@ public class CameraXProxy {
    * Converts a {@link ResolutionInfo} instance to a {@link Size} for setting the target resolution
    * of {@link UseCase}s.
    */
-  public static Size sizeFromResolution(@NonNull ResolutionInfo resolutionInfo) {
+  public static @NonNull Size sizeFromResolution(@NonNull ResolutionInfo resolutionInfo) {
     return new Size(resolutionInfo.getWidth().intValue(), resolutionInfo.getHeight().intValue());
   }
 
-  public CameraSelector.Builder createCameraSelectorBuilder() {
+  public @NonNull CameraSelector.Builder createCameraSelectorBuilder() {
     return new CameraSelector.Builder();
   }
 
-  public CameraPermissionsManager createCameraPermissionsManager() {
+  /** Creates an instance of {@link CameraPermissionsManager}. */
+  public @NonNull CameraPermissionsManager createCameraPermissionsManager() {
     return new CameraPermissionsManager();
   }
 
-  public DeviceOrientationManager createDeviceOrientationManager(
+  /** Creates an instance of the {@link DeviceOrientationManager}. */
+  public @NonNull DeviceOrientationManager createDeviceOrientationManager(
       @NonNull Activity activity,
       @NonNull Boolean isFrontFacing,
-      @NonNull int sensorOrientation,
+      int sensorOrientation,
       @NonNull DeviceOrientationManager.DeviceOrientationChangeCallback callback) {
     return new DeviceOrientationManager(activity, isFrontFacing, sensorOrientation, callback);
   }
 
-  public Preview.Builder createPreviewBuilder() {
+  /** Creates a builder for an instance of the {@link Preview} use case. */
+  public @NonNull Preview.Builder createPreviewBuilder() {
     return new Preview.Builder();
   }
 
-  public Surface createSurface(@NonNull SurfaceTexture surfaceTexture) {
+  /** Creates a {@link Surface} instance from the specified {@link SurfaceTexture}. */
+  public @NonNull Surface createSurface(@NonNull SurfaceTexture surfaceTexture) {
     return new Surface(surfaceTexture);
   }
 
   /**
-   * Creates an instance of the {@code SystemServicesFlutterApiImpl}.
+   * Creates an instance of the {@link SystemServicesFlutterApiImpl}.
    *
    * <p>Included in this class to utilize the callback methods it provides, e.g. {@code
    * onCameraError(String)}.
    */
-  public SystemServicesFlutterApiImpl createSystemServicesFlutterApiImpl(
+  public @NonNull SystemServicesFlutterApiImpl createSystemServicesFlutterApiImpl(
       @NonNull BinaryMessenger binaryMessenger) {
     return new SystemServicesFlutterApiImpl(binaryMessenger);
   }
 
+  /** Creates an instance of {@link Recorder.Builder}. */
+  @NonNull
+  public Recorder.Builder createRecorderBuilder() {
+    return new Recorder.Builder();
+  }
+
+  /** Creates a builder for an instance of the {@link ImageCapture} use case. */
+  @NonNull
   public ImageCapture.Builder createImageCaptureBuilder() {
     return new ImageCapture.Builder();
   }