+++ packages/camera/camera_android_camerax/pigeons/camerax_library.dart
@@ -46,6 +46,46 @@ class CameraPermissionsErrorData {
   String description;
 }
 
+/// The states the camera can be in.
+///
+/// See https://developer.android.com/reference/androidx/camera/core/CameraState.Type.
+enum CameraStateType {
+  closed,
+  closing,
+  open,
+  opening,
+  pendingOpen,
+}
+
+class CameraStateTypeData {
+  late CameraStateType value;
+}
+
+/// The types (T) properly wrapped to be used as a LiveData<T>.
+///
+/// If you need to add another type to support a type S to use a LiveData<S> in
+/// this plugin, ensure the following is done on the Dart side:
+///
+///  * In `../lib/src/live_data.dart`, add new cases for S in
+///    `_LiveDataHostApiImpl#getValueFromInstances` to get the current value of
+///    type S from a LiveData<S> instance and in `LiveDataFlutterApiImpl#create`
+///    to create the expected type of LiveData<S> when requested.
+///
+/// On the native side, ensure the following is done:
+///
+///  * Update `LiveDataHostApiImpl#getValue` is updated to properly return
+///    identifiers for instances of type S.
+///  * Update `ObserverFlutterApiWrapper#onChanged` to properly handle receiving
+///    calls with instances of type S if a LiveData<S> instance is observed.
+enum LiveDataSupportedType {
+  cameraState,
+  zoomState,
+}
+
+class LiveDataSupportedTypeData {
+  late LiveDataSupportedType value;
+}
+
 class ExposureCompensationRange {
   ExposureCompensationRange({
     required this.minCompensation,
@@ -77,6 +117,8 @@ abstract class JavaObjectFlutterApi {
 @HostApi(dartHostTestHandler: 'TestCameraInfoHostApi')
 abstract class CameraInfoHostApi {
   int getSensorRotationDegrees(int identifier);
+
+  int getCameraState(int identifier);
 
   int getExposureState(int identifier);
 
@@ -141,6 +183,8 @@ abstract class SystemServicesHostApi {
       bool isFrontFacing, int sensorOrientation);
 
   void stopListeningForDeviceOrientationChange();
+
+  String getTempFilePath(String prefix, String suffix);
 }
 
 @FlutterApi()
@@ -161,6 +205,60 @@ abstract class PreviewHostApi {
   ResolutionInfo getResolutionInfo(int identifier);
 }
 
+@HostApi(dartHostTestHandler: 'TestVideoCaptureHostApi')
+abstract class VideoCaptureHostApi {
+  int withOutput(int videoOutputId);
+
+  int getOutput(int identifier);
+}
+
+@FlutterApi()
+abstract class VideoCaptureFlutterApi {
+  void create(int identifier);
+}
+
+@HostApi(dartHostTestHandler: 'TestRecorderHostApi')
+abstract class RecorderHostApi {
+  void create(int identifier, int? aspectRatio, int? bitRate);
+
+  int getAspectRatio(int identifier);
+
+  int getTargetVideoEncodingBitRate(int identifier);
+
+  int prepareRecording(int identifier, String path);
+}
+
+@FlutterApi()
+abstract class RecorderFlutterApi {
+  void create(int identifier, int? aspectRatio, int? bitRate);
+}
+
+@HostApi(dartHostTestHandler: 'TestPendingRecordingHostApi')
+abstract class PendingRecordingHostApi {
+  int start(int identifier);
+}
+
+@FlutterApi()
+abstract class PendingRecordingFlutterApi {
+  void create(int identifier);
+}
+
+@HostApi(dartHostTestHandler: 'TestRecordingHostApi')
+abstract class RecordingHostApi {
+  void close(int identifier);
+
+  void pause(int identifier);
+
+  void resume(int identifier);
+
+  void stop(int identifier);
+}
+
+@FlutterApi()
+abstract class RecordingFlutterApi {
+  void create(int identifier);
+}
+
 @HostApi(dartHostTestHandler: 'TestImageCaptureHostApi')
 abstract class ImageCaptureHostApi {
   void create(int identifier, int? flashMode, ResolutionInfo? targetResolution);
@@ -169,6 +267,11 @@ abstract class ImageCaptureHostApi {
 
   @async
   String takePicture(int identifier);
+}
+
+@FlutterApi()
+abstract class CameraStateFlutterApi {
+  void create(int identifier, CameraStateTypeData type, int? errorIdentifier);
 }
 
 @FlutterApi()
@@ -196,6 +299,35 @@ abstract class ImageAnalysisHostApi {
 @HostApi(dartHostTestHandler: 'TestAnalyzerHostApi')
 abstract class AnalyzerHostApi {
   void create(int identifier);
+}
+
+@HostApi(dartHostTestHandler: 'TestObserverHostApi')
+abstract class ObserverHostApi {
+  void create(int identifier);
+}
+
+@FlutterApi()
+abstract class ObserverFlutterApi {
+  void onChanged(int identifier, int valueIdentifier);
+}
+
+@FlutterApi()
+abstract class CameraStateErrorFlutterApi {
+  void create(int identifier, int code);
+}
+
+@HostApi(dartHostTestHandler: 'TestLiveDataHostApi')
+abstract class LiveDataHostApi {
+  void observe(int identifier, int observerIdentifier);
+
+  void removeObservers(int identifier);
+
+  int? getValue(int identifier, LiveDataSupportedTypeData type);
+}
+
+@FlutterApi()
+abstract class LiveDataFlutterApi {
+  void create(int identifier, LiveDataSupportedTypeData type);
 }
 
 @FlutterApi()