+++ packages/file_selector/file_selector_macos/example/macos/RunnerTests/RunnerTests.swift
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import FlutterMacOS
+import UniformTypeIdentifiers
 import XCTest
 
 @testable import file_selector_macos
@@ -160,7 +161,7 @@ class exampleTests: XCTestCase {
       baseOptions: SavePanelOptions(
         allowedFileTypes: AllowedTypes(
           extensions: ["txt", "json"],
-          mimeTypes: [],
+          mimeTypes: ["text/html"],
           utis: ["public.text", "public.image"])))
     plugin.displayOpenPanel(options: options) { result in
       switch result {
@@ -175,7 +176,62 @@ class exampleTests: XCTestCase {
     wait(for: [called], timeout: 0.5)
     XCTAssertNotNil(panelController.openPanel)
     if let panel = panelController.openPanel {
+      if #available(macOS 11.0, *) {
+        XCTAssertTrue(panel.allowedContentTypes.contains(UTType.plainText))
+        XCTAssertTrue(panel.allowedContentTypes.contains(UTType.json))
+        XCTAssertTrue(panel.allowedContentTypes.contains(UTType.html))
+        XCTAssertTrue(panel.allowedContentTypes.contains(UTType.image))
+      } else {
+        // MIME type is not supported for the legacy codepath, but the rest should be set.
+        XCTAssertEqual(panel.allowedFileTypes, ["txt", "json", "public.text", "public.image"])
+      }
+    }
+  }
+
+  func testOpenWithFilterLegacy() throws {
+    let panelController = TestPanelController()
+    let plugin = FileSelectorPlugin(
+      viewProvider: TestViewProvider(),
+      panelController: panelController)
+    plugin.forceLegacyTypes = true
+
+    let returnPath = "/foo/bar"
+    panelController.openURLs = [URL(fileURLWithPath: returnPath)]
+
+    let called = XCTestExpectation()
+    let options = OpenPanelOptions(
+      allowsMultipleSelection: true,
+      canChooseDirectories: false,
+      canChooseFiles: true,
+      baseOptions: SavePanelOptions(
+        allowedFileTypes: AllowedTypes(
+          extensions: ["txt", "json"],
+          mimeTypes: ["text/html"],
+          utis: ["public.text", "public.image"])))
+    plugin.displayOpenPanel(options: options) { result in
+      switch result {
+      case .success(let paths):
+        XCTAssertEqual(paths[0], returnPath)
+      case .failure(let error):
+        XCTFail("\(error)")
+      }
+      called.fulfill()
+    }
+
+    wait(for: [called], timeout: 0.5)
+    XCTAssertNotNil(panelController.openPanel)
+    if let panel = panelController.openPanel {
+      // On the legacy path, the allowedFileTypes should be set directly.
       XCTAssertEqual(panel.allowedFileTypes, ["txt", "json", "public.text", "public.image"])
+
+      // They should also be translated to corresponding allowed content types.
+      if #available(macOS 11.0, *) {
+        XCTAssertTrue(panel.allowedContentTypes.contains(UTType.plainText))
+        XCTAssertTrue(panel.allowedContentTypes.contains(UTType.json))
+        XCTAssertTrue(panel.allowedContentTypes.contains(UTType.image))
+        // MIME type is not supported for the legacy codepath.
+        XCTAssertFalse(panel.allowedContentTypes.contains(UTType.html))
+      }
     }
   }