+++ packages/go_router_builder/example/lib/main.dart
@@ -4,6 +4,8 @@
 
 // ignore_for_file: public_member_api_docs
 
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
@@ -40,13 +42,13 @@ class App extends StatelessWidget {
     redirect: (BuildContext context, GoRouterState state) {
       final bool loggedIn = loginInfo.loggedIn;
 
-      // check just the subloc in case there are query parameters
+      // check just the matchedLocation in case there are query parameters
       final String loginLoc = const LoginRoute().location;
-      final bool goingToLogin = state.subloc == loginLoc;
+      final bool goingToLogin = state.matchedLocation == loginLoc;
 
       // the user is not logged in and not headed to /login, they need to login
       if (!loggedIn && !goingToLogin) {
-        return LoginRoute(fromPage: state.subloc).location;
+        return LoginRoute(fromPage: state.matchedLocation).location;
       }
 
       // the user is logged in and headed to /login, no need to login again
@@ -185,7 +187,7 @@ class HomeScreen extends StatelessWidget {
                   value: '2',
                   child: const Text('Push w/ return value'),
                   onTap: () async {
-                    FamilyCountRoute(familyData.length)
+                    unawaited(FamilyCountRoute(familyData.length)
                         .push<int>(context)
                         .then((int? value) {
                       if (value != null) {
@@ -195,7 +197,7 @@ class HomeScreen extends StatelessWidget {
                           ),
                         );
                       }
-                    });
+                    }));
                   },
                 ),
                 PopupMenuItem<String>(
@@ -257,7 +259,7 @@ class PersonScreen extends StatelessWidget {
               title: Text(
                   '${person.name} ${family.name} is ${person.age} years old'),
             ),
-            for (MapEntry<PersonDetails, String> entry
+            for (final MapEntry<PersonDetails, String> entry
                 in person.details.entries)
               ListTile(
                 title: Text(