+++ packages/local_auth/local_auth_ios/lib/local_auth_ios.dart
@@ -2,9 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/foundation.dart' show visibleForTesting;
 import 'package:flutter/services.dart';
 import 'package:local_auth_platform_interface/local_auth_platform_interface.dart';
 
+import 'src/messages.g.dart';
 import 'types/auth_messages_ios.dart';
 
 export 'package:local_auth_ios/types/auth_messages_ios.dart';
@@ -12,15 +14,19 @@ export 'package:local_auth_platform_interface/types/auth_messages.dart';
 export 'package:local_auth_platform_interface/types/auth_options.dart';
 export 'package:local_auth_platform_interface/types/biometric_type.dart';
 
-const MethodChannel _channel =
-    MethodChannel('plugins.flutter.io/local_auth_ios');
-
 /// The implementation of [LocalAuthPlatform] for iOS.
 class LocalAuthIOS extends LocalAuthPlatform {
+  /// Creates a new plugin implementation instance.
+  LocalAuthIOS({
+    @visibleForTesting LocalAuthApi? api,
+  }) : _api = api ?? LocalAuthApi();
+
   /// Registers this class as the default instance of [LocalAuthPlatform].
   static void registerWith() {
     LocalAuthPlatform.instance = LocalAuthIOS();
   }
+
+  final LocalAuthApi _api;
 
   @override
   Future<bool> authenticate({
@@ -29,58 +35,87 @@ class LocalAuthIOS extends LocalAuthPlatform {
     AuthenticationOptions options = const AuthenticationOptions(),
   }) async {
     assert(localizedReason.isNotEmpty);
-    final Map<String, Object> args = <String, Object>{
-      'localizedReason': localizedReason,
-      'useErrorDialogs': options.useErrorDialogs,
-      'stickyAuth': options.stickyAuth,
-      'sensitiveTransaction': options.sensitiveTransaction,
-      'biometricOnly': options.biometricOnly,
-    };
-    args.addAll(const IOSAuthMessages().args);
-    for (final AuthMessages messages in authMessages) {
-      if (messages is IOSAuthMessages) {
-        args.addAll(messages.args);
-      }
+    final AuthResultDetails resultDetails = await _api.authenticate(
+        AuthOptions(
+            biometricOnly: options.biometricOnly,
+            sticky: options.stickyAuth,
+            useErrorDialogs: options.useErrorDialogs),
+        _pigeonStringsFromAuthMessages(localizedReason, authMessages));
+    // TODO(stuartmorgan): Replace this with structured errors, coordinated
+    // across all platform implementations, per
+    // https://github.com/flutter/flutter/wiki/Contributing-to-Plugins-and-Packages#platform-exception-handling
+    // The PlatformExceptions thrown here are for compatibiilty with the
+    // previous Objective-C implementation.
+    switch (resultDetails.result) {
+      case AuthResult.success:
+        return true;
+      case AuthResult.failure:
+        return false;
+      case AuthResult.errorNotAvailable:
+        throw PlatformException(
+            code: 'NotAvailable',
+            message: resultDetails.errorMessage,
+            details: resultDetails.errorDetails);
+      case AuthResult.errorNotEnrolled:
+        throw PlatformException(
+            code: 'NotEnrolled',
+            message: resultDetails.errorMessage,
+            details: resultDetails.errorDetails);
+      case AuthResult.errorPasscodeNotSet:
+        throw PlatformException(
+            code: 'PasscodeNotSet',
+            message: resultDetails.errorMessage,
+            details: resultDetails.errorDetails);
     }
-    return (await _channel.invokeMethod<bool>('authenticate', args)) ?? false;
   }
 
   @override
   Future<bool> deviceSupportsBiometrics() async {
-    return (await _channel.invokeMethod<bool>('deviceSupportsBiometrics')) ??
-        false;
+    return _api.deviceCanSupportBiometrics();
   }
 
   @override
   Future<List<BiometricType>> getEnrolledBiometrics() async {
-    final List<String> result = (await _channel.invokeListMethod<String>(
-          'getEnrolledBiometrics',
-        )) ??
-        <String>[];
-    final List<BiometricType> biometrics = <BiometricType>[];
-    for (final String value in result) {
-      switch (value) {
-        case 'face':
-          biometrics.add(BiometricType.face);
-          break;
-        case 'fingerprint':
-          biometrics.add(BiometricType.fingerprint);
-          break;
-        case 'iris':
-          biometrics.add(BiometricType.iris);
-          break;
+    final List<AuthBiometricWrapper?> result =
+        await _api.getEnrolledBiometrics();
+    return result
+        .cast<AuthBiometricWrapper>()
+        .map((AuthBiometricWrapper entry) {
+      switch (entry.value) {
+        case AuthBiometric.face:
+          return BiometricType.face;
+        case AuthBiometric.fingerprint:
+          return BiometricType.fingerprint;
       }
-    }
-    return biometrics;
+    }).toList();
   }
 
   @override
-  Future<bool> isDeviceSupported() async =>
-      (await _channel.invokeMethod<bool>('isDeviceSupported')) ?? false;
+  Future<bool> isDeviceSupported() async => _api.isDeviceSupported();
 
   /// Always returns false as this method is not supported on iOS.
   @override
-  Future<bool> stopAuthentication() async {
-    return false;
+  Future<bool> stopAuthentication() async => false;
+
+  AuthStrings _pigeonStringsFromAuthMessages(
+      String localizedReason, Iterable<AuthMessages> messagesList) {
+    IOSAuthMessages? messages;
+    for (final AuthMessages entry in messagesList) {
+      if (entry is IOSAuthMessages) {
+        messages = entry;
+        break;
+      }
+    }
+    return AuthStrings(
+      reason: localizedReason,
+      lockOut: messages?.lockOut ?? iOSLockOut,
+      goToSettingsButton: messages?.goToSettingsButton ?? goToSettings,
+      goToSettingsDescription:
+          messages?.goToSettingsDescription ?? iOSGoToSettingsDescription,
+      // TODO(stuartmorgan): The default's name is confusing here for legacy
+      // reasons; this should be fixed as part of some future breaking change.
+      cancelButton: messages?.cancelButton ?? iOSOkButton,
+      localizedFallbackTitle: messages?.localizedFallbackTitle,
+    );
   }
 }