+++ packages/file_selector/file_selector_linux/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.3/Dart 2.18.
+
 ## 0.9.1+3
 
 * Sets a cmake_policy compatibility version to fix build warnings.