+++ packages/local_auth/local_auth_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.3/Dart 2.18.
+
+## 1.0.28
+
+* Removes unused resources as indicated by Android lint warnings.
+
 ## 1.0.27
 
 * Fixes compatibility with AGP versions older than 4.2.