+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,6 +1,15 @@
-## 3.7.0
+## 3.8.0
 
 * Adds support for fullscreen video.
+
+## 3.7.1
+
+* Removes obsolete null checks on non-nullable values.
+
+## 3.7.0
+
+* Adds support to accept third party cookies. See
+  `AndroidWebViewCookieManager.setAcceptThirdPartyCookies`.
 
 ## 3.6.3